requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
pyarrow>=14.0.0
plotly>=5.20.0
Pillow>=10.0.0
numpy>=1.26.0
scipy>=1.12.0
